+++ install/setup_unattended.sh
@@ -37,15 +37,14 @@ esac
 
 printf "Checking Python version...\n"
 if hash python3 2>/dev/null; then
-  if ! python3 "${INSTALL_DIRECTORY}"/mycodo/scripts/upgrade_check.py --min_python_version "3.6"; then
-    printf "Incorrect Python version found. Mycodo requires Python >= 3.6.\n"
-    printf "If you're running Debian Stretch with Python 3.5, you will need to install at least Debian Buster with Python 3.7 to install the latest version of Mycodo.\n"
+  if ! python3 "${INSTALL_DIRECTORY}"/mycodo/scripts/upgrade_check.py --min_python_version "3.8"; then
+    printf "Error: Incorrect Python version found. Mycodo requires Python >= 3.8.\n"
     exit 1
   else
-    printf "Python >= 3.6 found. Continuing with the install.\n"
+    printf "Python >= 3.8 found. Continuing with the install.\n"
   fi
 else
-  printf "\npython3 was not found. Cannot proceed with the install without python3 (Python >= 3.6).\n"
+  printf "\nError: python3 binary required in PATH to proceed with the install.\n"
   exit 1
 fi